Finalize Travix CI integration
