+++ .travis.yml
@@ -1,9 +1,10 @@
 language: python
 python:
   - "2.7"
+  - "3.2"
   - "3.4"
   - "3.5"
 # command to install dependencies
-install: "pip install -r requirements.txt"
+install: pip install --user -r requirements.txt
 # command to run tests
 script: python setup.py test
